+++ flask_bluelogin/models/user.py
@@ -4,7 +4,7 @@ from .users import Users, AlreadyExistUserError
 
 
 class User(Model):
-    def __init__(self, id=None, password=None, groups=[], properties={}):
+    def __init__(self, id=None, password=None, groups=[], properties={}, active=True):
         """
         User 
 
@@ -16,11 +16,14 @@ class User(Model):
         :type properties: dict
         :param password: password of this User.
         :type password: str
+        :param active: state of this User.
+        :type active: boolean
         """
         self._id = id
         self._password = password
         self._groups = groups
         self._properties = properties
+        self._active = active
 
     @property
     def id(self):
@@ -72,7 +75,7 @@ class User(Model):
     def properties(self):
         """
         Gets dict of properties of this User.
-        list of group of user
+        dict of properties of user
 
         :return: The dict of properties of this User.
         :rtype: dict
@@ -89,6 +92,29 @@ class User(Model):
         :type properties: dict
         """
         self._properties= properties    
+
+    @property
+    def active(self):
+        """
+        Gets state of this User.
+        state of user
+
+        :return: The state of this User.
+        :rtype: boolean
+        """
+        return self._active
+
+    @active.setter
+    def active(self, active):
+        """
+        Sets state of this User.
+        state of user
+
+        :param active: The state of this User.
+        :type active: boolean
+        """
+        self._active= active   
+
 
     @property
     def password(self):
@@ -121,7 +147,7 @@ class User(Model):
     
     @property
     def is_active(self):
-            return True
+            return self.active
     
     @property
     def is_anonymous(self):