+++ CHANGES.rst
@@ -9,6 +9,7 @@ V. 0.2.0
 - add function add_group in User
 - add functionnality Users as generic
 - add function save in User
+- add property active for user
 
 V. 0.1.0
 ========